Create src folder and adjust the test to fit the architecture
