--- backend/auth/exception.py
@@ -1,6 +1,0 @@
-class EmailAlreadyRegisteredError(RuntimeError):
-    pass
-
-
-class UserNotFoundError(RuntimeError):
-    pass
