+++ backend/src/auth/exception.py
@@ -1,0 +1,6 @@
+class EmailAlreadyRegisteredError(RuntimeError):
+    pass
+
+
+class UserNotFoundError(RuntimeError):
+    pass
