+++ backend/src/auth/util.py
@@ -13,9 +13,9 @@ import jwt
 from flask import current_app, make_response, redirect, request
 from sqlalchemy import select
 
-from auth.exception import EmailAlreadyRegisteredError, UserNotFoundError
-from database import db
-from models import User
+from src.auth.exception import EmailAlreadyRegisteredError, UserNotFoundError
+from src.database import db
+from src.models import User
 from util import SingleMessageStatus
 
 if TYPE_CHECKING:
