+++ backend/src/static/exception.py
@@ -1,0 +1,2 @@
+class ImageNotExistError(FileNotFoundError):
+    pass
