+++ backend/tests/unit_tests/auth/test_auth.py
@@ -6,9 +6,9 @@ from typing import TYPE_CHECKING, Any
 
 import pytest
 
-from auth.util import Gender, HS256JWTCodec
-from database import db
-from models import User
+from src.auth.util import Gender, HS256JWTCodec
+from src.database import db
+from src.models import User
 from tests.util import assert_not_raise
 
 if TYPE_CHECKING:
