+++ backend/tests/unit_tests/conftest.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Generator
 import pytest
 
 from app import create_app
-from database import create_db, db
+from src.database import create_db, db
 from tests.util import executescript
 
 if TYPE_CHECKING:
